+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/ProcessResourceImpl.java
@@ -13,6 +13,7 @@ import com.walmartlabs.concord.server.process.pipelines.ProjectPipeline;
 import com.walmartlabs.concord.server.process.pipelines.ResumePipeline;
 import com.walmartlabs.concord.server.process.pipelines.processors.Chain;
 import com.walmartlabs.concord.server.process.queue.ProcessQueueDao;
+import com.walmartlabs.concord.server.process.state.ProcessStateManager;
 import com.walmartlabs.concord.server.project.ProjectDao;
 import com.walmartlabs.concord.server.security.UserPrincipal;
 import io.takari.bpm.api.ExecutionException;
@@ -36,12 +37,13 @@ import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.StreamingOutput;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
+
+import static com.walmartlabs.concord.server.process.state.ProcessStateManager.path;
 
 @Named
 public class ProcessResourceImpl implements ProcessResource, Resource {
@@ -54,6 +56,7 @@ public class ProcessResourceImpl implements ProcessResource, Resource {
     private final Chain projectPipeline;
     private final Chain resumePipeline;
     private final PayloadManager payloadManager;
+    private final ProcessStateManager stateManager;
     private final AgentManager agentManager;
     private final ConcordFormService formService;
 
@@ -64,6 +67,7 @@ public class ProcessResourceImpl implements ProcessResource, Resource {
                                ProjectPipeline projectPipeline,
                                ResumePipeline resumePipeline,
                                PayloadManager payloadManager,
+                               ProcessStateManager stateManager,
                                AgentManager agentManager,
                                ConcordFormService concordFormService) {
 
@@ -73,6 +77,7 @@ public class ProcessResourceImpl implements ProcessResource, Resource {
         this.projectPipeline = projectPipeline;
         this.resumePipeline = resumePipeline;
         this.payloadManager = payloadManager;
+        this.stateManager = stateManager;
         this.agentManager = agentManager;
         this.formService = concordFormService;
     }
@@ -275,18 +280,32 @@ public class ProcessResourceImpl implements ProcessResource, Resource {
     @Validate
     @WithTimer
     public Response downloadAttachment(String instanceId, String attachmentName) {
+        // TODO replace with javax.validation
         if (attachmentName.endsWith("/")) {
             throw new WebApplicationException("Invalid attachment name: " + attachmentName, Status.BAD_REQUEST);
         }
 
-        // TODO sanitize
-        Path p = payloadManager.getResource(instanceId, Constants.Files.JOB_ATTACHMENTS_DIR_NAME + "/" + attachmentName);
-        if (p == null) {
+        String resource = path(Constants.Files.JOB_ATTACHMENTS_DIR_NAME, attachmentName);
+        Optional<Path> o = stateManager.get(instanceId, resource, src -> {
+            try {
+                Path tmp = Files.createTempFile("attachment", ".bin");
+                try (OutputStream dst = Files.newOutputStream(tmp)) {
+                    IOUtils.copy(src, dst);
+                }
+                return Optional.of(tmp);
+            } catch (IOException e) {
+                throw new WebApplicationException("Error while exporting an attachment: " + attachmentName, e);
+            }
+        });
+
+        if (!o.isPresent()) {
             return Response.status(Status.NOT_FOUND).build();
         }
 
+        Path tmp = o.get();
+
         return Response.ok((StreamingOutput) out -> {
-            try (InputStream in = Files.newInputStream(p)) {
+            try (InputStream in = Files.newInputStream(tmp)) {
                 IOUtils.copy(in, out);
             }
         }).build();
@@ -310,15 +329,20 @@ public class ProcessResourceImpl implements ProcessResource, Resource {
 
     @SuppressWarnings("unchecked")
     private Map<String, Object> readArgs(String instanceId) {
-        Path p = payloadManager.getResource(instanceId, Constants.Files.REQUEST_DATA_FILE_NAME);
+        String resource = Constants.Files.REQUEST_DATA_FILE_NAME;
+        Optional<Map<String, Object>> o = stateManager.get(instanceId, resource, in -> {
+            try {
+                ObjectMapper om = new ObjectMapper();
 
-        try (InputStream in = Files.newInputStream(p)) {
-            ObjectMapper om = new ObjectMapper();
-            Map<String, Object> cfg = om.readValue(in, Map.class);
-            return (Map<String, Object>) cfg.get(Constants.Request.ARGUMENTS_KEY);
-        } catch (IOException e) {
-            throw new WebApplicationException("Error while reading request data", e);
-        }
+                Map<String, Object> cfg = om.readValue(in, Map.class);
+                Map<String, Object> args = (Map<String, Object>) cfg.get(Constants.Request.ARGUMENTS_KEY);
+
+                return Optional.ofNullable(args);
+            } catch (IOException e) {
+                throw new WebApplicationException("Error while reading request data", e);
+            }
+        });
+        return o.orElse(Collections.emptyMap());
     }
 
     private void process(String instanceId, Map<String, Object> params) {