+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ResumePipeline.java
@@ -15,6 +15,7 @@ public class ResumePipeline extends Pipeline {
                 LogFileProcessor.class,
                 ResumeStateStoringProcessor.class,
                 RequestDataStoringProcessor.class,
+                StateImportingProcessor.class,
                 EnqueueingResumeProcessor.class);
     }
 }