+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - concord-console: handle `ENQUEUED` status in the default form
 wizard;
+- concord-server: store process state in the DB;
 - concord-server: create the log file of a process as early as
 possible to log startup errors.